Fix issues reported by Codeclimate
