+++ hailstorm-web-client/src/DataCenterForm/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "DataCenterForm.tsx"
+}
